requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
requests
python-dotenv
openai